Back to editor now: switching things over to uint because who needs signed numbers anyways
FossilOrigin-Name: e70815ea6ed18e764fb350d3031560224ad32c889482ac0d0cef0281354fd07b

+++ editor/display.go
@@ -21,8 +21,8 @@ func DrawString(s tcell.Screen, x int, y int, stringToDraw string) {
 func (E *Editor) DrawEditorChars(xPos int, yPos int, leftChar rune) {
 	w, h := E.s.Size()
 	for rowindx := 0; rowindx < E.numrows && rowindx < h; rowindx++ {
-		for charindx := 0; charindx < E.rows[rowindx].length && charindx < w; charindx++ {
-			E.s.SetContent(charindx+2, rowindx, E.rows[rowindx].chars[charindx], nil, tcell.StyleDefault)
+		for charindx := uint(0); charindx < E.rows[rowindx].length && charindx < uint(w); charindx++ {
+			E.s.SetContent(int(charindx+2), rowindx, E.rows[rowindx].chars[charindx], nil, tcell.StyleDefault)
 		}
 		// lineNo := fmt.Sprintf("%d", rowindx)
 		E.s.SetContent(0, rowindx, LEFTSIDE_CHAR, nil, tcell.StyleDefault)
